CLOUDSYNC-26: Pre-emptively Authenticate for Fedora
When making requests to Fedora, CloudSync now provides credentials
for each request, whether challenged or not. This is necessary in order
to get copy tasks to work properly against Fedora 3.5
(and probably 3.6) because for managed datastream content GET requests
(and possibly others), Fedora doesn't both to send an AuthN challenge.

https://jira.duraspace.org/browse/CLOUDSYNC-26